image upload is updated

+++ middleware/imageupload.js
@@ -17,24 +17,32 @@ var imageupload=(request,response,next)=>{
         accessKeyId: env.AWS_ACCESS_KEY_ID,                           //AWS Access key  
         secretAccessKey: env.AWS_SECRET_ACCESS_KEY                     //AWS Screat access key
     });
-    var myfile = request.files.Image;
-    filename=request.files.Image.name;
-    var filename=filename.replace(/\s/g,'');                        //replacing white spaces from the user filename
-    myfile.originalFilename = Date.now() + filename;
-    const awsurl = "https://s3.amazonaws.com/"+BUCKET_NAME+'/';        //AWS url plus the bucket name a user upload an image        
-    var stream = fs.createReadStream(myfile.path);
-    s3fsImpl.writeFile(myfile.originalFilename, stream).then(() => {
-        fs.unlink(myfile.path, (err) => {
-            if (err) {
-                console.log(err);
-            }
-            else {
-                console.log("upload successfully");
-                request.imageurl = awsurl + myfile.originalFilename;      //returning the url to the user database 
-                next();
-            }
+    if(request.files.Image!=null)
+    {
+        var myfile = request.files.Image;
+        filename = request.files.Image.name;
+        var filename = filename.replace(/\s/g, '');                        //replacing white spaces from the user filename
+        myfile.originalFilename = Date.now() + filename;
+        const awsurl = "https://s3.amazonaws.com/" + BUCKET_NAME + '/';        //AWS url plus the bucket name a user upload an image        
+        var stream = fs.createReadStream(myfile.path);
+        s3fsImpl.writeFile(myfile.originalFilename, stream).then(() => {
+            fs.unlink(myfile.path, (err) => {
+                if (err) {
+                    console.log(err);
+                }
+                else {
+                    console.log("upload successfully");
+                    request.imageurl = awsurl + myfile.originalFilename;      //returning the url to the user database 
+                    next();
+                }
+            });
         });
-    });
+    }else{
+        next();
+    }
+
+
+    
 };
